Update CommitResolutionDialog #4712
