+++ ui/app/src/components/CommitResolutionDialog/index.ts
@@ -17,3 +17,5 @@
 export { default } from './CommitResolutionDialog';
 
 export * from './CommitResolutionDialog';
+
+export * from  './CommitResolutionDialogContainer';
